Move pip upgrade into the venv
In af6d45f028f61990ad0c9a4c448cd6504d900485 we implemented changes
to ensure that pip was the right version to be able to install
the kibana-selenium tests, but unfortunately the pip install was
executed outside of the venv, causing the objective not to be
achieved and other breakage.

In this patch we move the pip install/upgrade into the venv, where
it should have been implemented from the start.

JIRA: RI-482