Allow unbound variables for virtualenv activation
Enabling unbound variables appears to be required for
virtualenv activation. We also remove the verbose output
because it makes the output really difficult to read
because it shows the enture script. Instead we enable
debug output, as is done in most other scripts, because
it shows the commands run instead which is more useful.

JIRA: RI-492