ASC-1042 Enable multiple system actions (#3318)
This commit modifies the match for the 'system' gating action. Any
action that begins with the string 'system' will be matched. This is
done to allow multiple variants of the system action to be defined in CI
in order to segment jobs.

(cherry picked from commit 065a7c350a27ca51ccaf0807a9e34ea5cddec11e)