ASC-1042 Enable multiple system actions (#3318)
This commit modifies the match for the 'system' gating action. Any
action that begins with the string 'system' will be matched. This is
done to allow multiple variants of the system action to be defined in CI
in order to segment jobs.

(cherry picked from commit 065a7c350a27ca51ccaf0807a9e34ea5cddec11e)

The following conflict was resolved while committing this cherry pick:
```
41 <<<<<<< HEAD
42 if [[ $RE_JOB_ACTION == "system" ]]; then
43 =======
44
45 if [[ $RE_JOB_ACTION == system* ]]; then
46 >>>>>>> 065a7c35... ASC-1042 Enable multiple system actions (#3318)
```