[RI-538 and RI-561] ELK Fixes
This makes the ELK HAproxy services conditional based on whether or not
the proper inventory exists, if not those services are not enabled in
HAProxy.

JIRA: RI-538

The elk-deployment.yml file was renamed to deployment-elk.yml but the
references to elk-deployment.yml were not adjusted.  This rectifies
that.

JIRA: RI-561
(cherry picked from commit 78cea6ffccfbb2a0674b5f14a4504bf3199b4cca)