RO-4194 Update pike third-party dependencies
This change is the result of running gating/update_dependencies/*
  See update_dependencies in https://rpc-openstack.atlassian.net/wiki/spaces/RE/pages/19005457/RE+for+Projects for more details.

This change pulls in a new version of OpenStack-Ansible to fix the
failure introduced by the Ansible role ansible-sshd removing a number of
its tags. This issue also impacts rpc-differ and so this change also
updates rpc-differ to allow the removed tag to be mapped to its
replacement.