Add haproxy back-end check as part of AIO test
When building an environment, there should be no situation
where the resulting build has a haproxy back-end that is down.
Recently some implementations were merged which did this,
causing the use of snapshot AIO images to fail due to the
failed back-ends in haproxy.

To prevent this happening again, we ensure that part of the
AIO PR/PM test is to do the same check and to fail if there
is a back-end down. To ensure that the requirements for the
test are in place, we also set the test to install the package
it requires.

JIRA: RI-538
(cherry picked from commit 7348ce04dd6ec5910f952a8b2d3d04ba26523e7a)