Add new official grafana repo and rcbops aptly role
As of 7 January 2019, packagecloud is no longer used
as a repo for grafana. This adds a new mirror for it
and stops updates to the old mirror. We also include
the new mirror in the snapshots and remove the old
one.

Previously, we were using JP's fork of the aptly repo
to deploy and configure aptly. To ensure that we are
in control of the role repo, we have forked it to
rcbops and updated the configuration to use our fork.

ref: https://grafana.com/blog/2019/01/05/moving-to-packages.grafana.com/