ASC-978 Update "rpc-zigzag" Version
Update the script to pin to the 1.0 series for "rpc-zigzag". Also, added a
variable for the "rpc-zigzag" PyPI package at the top so it can be easily
updated in the future without hunting around in the script.

The ZigZag config file is contained as a heredoc within the script which is
written out to disk at runtime.